chapter-05/04-sessionization/01-incremental-delta-lake-spark/requirements.txt
pyspark==4.0.0
delta-spark==4.0.0
pandas==3.0.0
pyarrow==23.0.0